Attack Update
- Restructured the Attack System
- Attack Manager and Network Attack Manager are the only two scripts now
- Removed all other useless scripts for attacking
