+++ Seize Command/Assets/Prefabs/Player.prefab
@@ -97,7 +97,7 @@ GameObject:
   - component: {fileID: 4964248940627659643}
   - component: {fileID: -8914290158638435365}
   - component: {fileID: 1414329148677328424}
-  - component: {fileID: 654514585087359453}
+  - component: {fileID: 4119761002428496995}
   - component: {fileID: 5669298925497299681}
   m_Layer: 8
   m_Name: Player
@@ -233,7 +233,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 0
---- !u!114 &654514585087359453
+--- !u!114 &4119761002428496995
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -242,11 +242,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 2536958507248721030}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3dda5d1e854d9ab4092262556ccfe5f9, type: 3}
+  m_Script: {fileID: 11500000, guid: 7571b6a1f10f05d4bb8a190d36ed5167, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   gun: {fileID: 3085805901309814881}
-  networkIdentity: {fileID: 654283526}
 --- !u!114 &5669298925497299681
 MonoBehaviour:
   m_ObjectHideFlags: 0
