--- Seize Command/Assets/Scripts/Attack/AbstractNetworkAttack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bafb4cdcf83e7be4d9a17cdf9fa7f991
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
