+++ Seize Command/Assets/Scripts/Attack/AttackManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3dda5d1e854d9ab4092262556ccfe5f9
+guid: 4c83dd73f380003479cd5618d9c3a2b4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
