+++ Seize Command/Assets/Scripts/Attack/NetworkAttackManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d9cf9715319779745944b872289bf1c1
+guid: 7571b6a1f10f05d4bb8a190d36ed5167
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
